requirements.txt
django
psycopg2-binary
python-dotenv
django-jazzmin
Pillow
django-filter
pandas
openpyxl
